refactor(streaming): extract process/subtitle/cache trio from hls_service
